+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.1.2 under development
 
-- no changes in this release.
+- Enh #73: Set `en_US` as default locale for translator (@vjik)
 
 ## 1.1.1 September 09, 2022