+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.2 under development
 
+- Enh #70: Add `Stringable` type support for `$id` argument in `Translator::translate()` (@xepozz)
 - Enh #74: Dispatch `MissingTranslationCategoryEvent` once per category (@vjik)
 - Enh #73: Set `en_US` as default locale for translator (@vjik)
 - Enh #72, #75: Format messages when missing translation category (@vjik)